move test folder out of src
