+++ tsconfig.json
@@ -16,5 +16,5 @@
     "noFallthroughCasesInSwitch": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["./src", "./examples"]
+  "include": ["./src", "./examples", "./test"]
 }
